+++ src/implementations/Benqi/BenqiMultiStake.sol
@@ -196,7 +196,7 @@ contract BenqiMultiStake is StrategyV5Chainlink {
      * @notice Returns the investment in asset asset for the specified input
      * @return total Amount invested
      */
-    function invested(uint8 _index) public view override returns (uint256) {
+    function invested(uint256 _index) public view override returns (uint256) {
         return _inputToAsset(investedInput(_index), _index);
     }
 
@@ -205,7 +205,7 @@ contract BenqiMultiStake is StrategyV5Chainlink {
      * @return total Amount invested
      */
     function investedInput(
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _stakedInput(_index);
     }
@@ -216,7 +216,7 @@ contract BenqiMultiStake is StrategyV5Chainlink {
      */
     function _stakeToInput(
         uint256 _amount,
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _amount.mulDiv(
             qiTokens[_index].exchangeRateStored(),
@@ -229,7 +229,7 @@ contract BenqiMultiStake is StrategyV5Chainlink {
      */
     function _inputToStake(
         uint256 _amount,
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _amount.mulDiv(
             1e18,
@@ -241,7 +241,7 @@ contract BenqiMultiStake is StrategyV5Chainlink {
      * @return Input value of the LP/staked balance
      */
     function _stakedInput(
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _stakeToInput(qiTokens[_index].balanceOf(address(this)), _index);
     }
