+++ src/implementations/Hop/HopMultiStake.sol
@@ -250,7 +250,7 @@ contract HopMultiStake is StrategyV5Chainlink {
      * @notice Returns the investment in asset asset for the specified input
      * @return total Amount invested
      */
-    function invested(uint8 _index) public view override returns (uint256) {
+    function invested(uint256 _index) public view override returns (uint256) {
         return
             _stakeToAsset(
                 rewardPools[_index][0].balanceOf(address(this)),
@@ -263,7 +263,7 @@ contract HopMultiStake is StrategyV5Chainlink {
      * @return total Amount invested
      */
     function investedInput(
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return
             _stakeToInput(
@@ -278,7 +278,7 @@ contract HopMultiStake is StrategyV5Chainlink {
      */
     function _stakeToInput(
         uint256 _amount,
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return
             _amount.mulDiv(
@@ -293,7 +293,7 @@ contract HopMultiStake is StrategyV5Chainlink {
      */
     function _inputToStake(
         uint256 _amount,
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return
             _amount.mulDiv(
@@ -307,7 +307,7 @@ contract HopMultiStake is StrategyV5Chainlink {
      * @return Input value of the LP/staked balance
      */
     function _stakedInput(
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return
             _stakeToInput(
