+++ src/implementations/Moonwell/MoonwellMultiStake.sol
@@ -192,7 +192,7 @@ contract MoonwellMultiStake is StrategyV5Chainlink {
      * @notice Returns the investment in asset asset for the specified input
      * @return total Amount invested
      */
-    function invested(uint8 _index) public view override returns (uint256) {
+    function invested(uint256 _index) public view override returns (uint256) {
         return _inputToAsset(investedInput(_index), _index);
     }
 
@@ -201,7 +201,7 @@ contract MoonwellMultiStake is StrategyV5Chainlink {
      * @return total Amount invested
      */
     function investedInput(
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _stakedInput(_index);
     }
@@ -212,7 +212,7 @@ contract MoonwellMultiStake is StrategyV5Chainlink {
      */
     function _stakeToInput(
         uint256 _amount,
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _amount.mulDiv(mTokens[_index].exchangeRateStored(), 1e18);
     }
@@ -223,7 +223,7 @@ contract MoonwellMultiStake is StrategyV5Chainlink {
      */
     function _inputToStake(
         uint256 _amount,
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _amount.mulDiv(1e18, mTokens[_index].exchangeRateStored());
     }
@@ -233,7 +233,7 @@ contract MoonwellMultiStake is StrategyV5Chainlink {
      * @return Input value of the LP/staked balance
      */
     function _stakedInput(
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _stakeToInput(mTokens[_index].balanceOf(address(this)), _index);
     }
