+++ src/implementations/Toros/TorosMultiStake.sol
@@ -190,7 +190,7 @@ contract TorosMultiStake is StrategyV5Chainlink {
      * @notice Returns the investment in asset asset for the specified input
      * @return total Amount invested
      */
-    function invested(uint8 _index) public view override returns (uint256) {
+    function invested(uint256 _index) public view override returns (uint256) {
         return _inputToAsset(investedInput(_index), _index);
     }
 
@@ -199,7 +199,7 @@ contract TorosMultiStake is StrategyV5Chainlink {
      * @return total Amount invested
      */
     function investedInput(
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _stakedInput(_index);
     }
@@ -210,7 +210,7 @@ contract TorosMultiStake is StrategyV5Chainlink {
      */
     function _stakeToInput(
         uint256 _amount,
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _usdToInput(_amount.mulDiv(pools[_index].tokenPrice(), 1e12), _index);
     }
@@ -221,7 +221,7 @@ contract TorosMultiStake is StrategyV5Chainlink {
      */
     function _inputToStake(
         uint256 _amount,
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _inputToUsd(_amount, _index).mulDiv(1e12 * poolDecimals[_index], pools[_index].tokenPrice()); // eg. 1e6+1e12+1e18-1e18 = 1e18
     }
@@ -231,7 +231,7 @@ contract TorosMultiStake is StrategyV5Chainlink {
      * @return Input value of the LP/staked balance
      */
     function _stakedInput(
-        uint8 _index
+        uint256 _index
     ) internal view override returns (uint256) {
         return _stakeToInput(pools[_index].balanceOf(address(this)), _index);
     }
